Technically, it makes more sense to disallow this as an option, but I feel like this could be useful as a 'neat trick' to reduce storage space usage in real time, so I'm keeping it around, it's just hidden in a command line argument that has to be set.

+++ gobii-process/src/main/java/org/gobiiproject/gobiiprocess/digester/GobiiFileReader.java
@@ -444,7 +444,9 @@ public class GobiiFileReader {
 						ErrorLogger.logError("FileReader","Error in table "+key);
 					}
 					else{
-						deleteIFLFiles(dstDir,key);
+						if(LoaderGlobalConfigs.getDeleteIntermediateFiles()){
+							deleteIFLFiles(dstDir,key);
+						}
 					}
 
 				}
@@ -880,9 +882,6 @@ public class GobiiFileReader {
 	 * @param tableName
 	 */
 	private static void deleteIFLFiles(File directory, String tableName) {
-		if(LoaderGlobalConfigs.getKeepIntermediateFiles()){
-			return;
-		}
 		File[] fileList = directory.listFiles();
 		if (fileList == null) return;
 		for (File f : fileList) {
